Reorganize handleSyncOp
