+++ pkg/pillar/cmd/downloader/dns.go
@@ -1,0 +1,42 @@
+package downloader
+
+import (
+	"github.com/google/go-cmp/cmp"
+	"github.com/lf-edge/eve/pkg/pillar/cast"
+	"github.com/lf-edge/eve/pkg/pillar/types"
+	log "github.com/sirupsen/logrus"
+)
+
+// Handles both create and modify events
+func handleDNSModify(ctxArg interface{}, key string, statusArg interface{}) {
+
+	ctx := ctxArg.(*downloaderContext)
+	status := cast.CastDeviceNetworkStatus(statusArg)
+	if key != "global" {
+		log.Infof("handleDNSModify: ignoring %s\n", key)
+		return
+	}
+	log.Infof("handleDNSModify for %s\n", key)
+	if cmp.Equal(ctx.deviceNetworkStatus, status) {
+		log.Infof("handleDNSModify unchanged\n")
+		return
+	}
+	ctx.deviceNetworkStatus = status
+	log.Infof("handleDNSModify %d free management ports addresses; %d any\n",
+		types.CountLocalAddrFreeNoLinkLocal(ctx.deviceNetworkStatus),
+		types.CountLocalAddrAnyNoLinkLocal(ctx.deviceNetworkStatus))
+
+	log.Infof("handleDNSModify done for %s\n", key)
+}
+
+func handleDNSDelete(ctxArg interface{}, key string, statusArg interface{}) {
+
+	ctx := ctxArg.(*downloaderContext)
+	log.Infof("handleDNSDelete for %s\n", key)
+	if key != "global" {
+		log.Infof("handleDNSDelete: ignoring %s\n", key)
+		return
+	}
+	ctx.deviceNetworkStatus = types.DeviceNetworkStatus{}
+	log.Infof("handleDNSDelete done for %s\n", key)
+}
