+++ pkg/pillar/cmd/downloader/downloader.go
@@ -8,21 +8,13 @@
 package downloader
 
 import (
-	"encoding/json"
-	"errors"
 	"flag"
 	"fmt"
-	"io/ioutil"
-	"net"
-	"net/url"
 	"os"
-	"regexp"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/google/go-cmp/cmp"
-	zconfig "github.com/lf-edge/eve/api/go/config"
 	"github.com/lf-edge/eve/pkg/pillar/agentlog"
 	"github.com/lf-edge/eve/pkg/pillar/cast"
 	"github.com/lf-edge/eve/pkg/pillar/diskmetrics"
@@ -30,18 +22,14 @@ import (
 	"github.com/lf-edge/eve/pkg/pillar/pidfile"
 	"github.com/lf-edge/eve/pkg/pillar/pubsub"
 	"github.com/lf-edge/eve/pkg/pillar/types"
-	"github.com/lf-edge/eve/pkg/pillar/wrap"
 	"github.com/lf-edge/eve/pkg/pillar/zedUpload"
 	"github.com/lf-edge/eve/pkg/pillar/zedcloud"
-	"github.com/satori/go.uuid"
+	uuid "github.com/satori/go.uuid"
 	log "github.com/sirupsen/logrus"
 )
 
 const (
 	agentName = "downloader"
-	// persistRktLocalConfigBase - Base Dir used for LocalConfigDir for
-	//  rkt Container images
-	persistRktLocalConfigBase = types.PersistDir + "/rktlocal"
 )
 
 // Go doesn't like this as a constant
@@ -1090,326 +1078,6 @@ func downloaderSubscription(ctx *downloaderContext, objType string) *pubsub.Subs
 }
 
 // cloud storage interface functions/APIs
-
-func doHttp(ctx *downloaderContext, status *types.DownloaderStatus,
-	syncOp zedUpload.SyncOpType, serverUrl, dpath string, maxsize uint64,
-	ifname string, ipSrc net.IP, filename, locFilename string) error {
-
-	auth := &zedUpload.AuthInput{
-		AuthType: "http",
-	}
-
-	trType := zedUpload.SyncHttpTr
-
-	// create Endpoint
-	dEndPoint, err := ctx.dCtx.NewSyncerDest(trType, serverUrl, dpath, auth)
-	if err != nil {
-		log.Errorf("NewSyncerDest failed: %s\n", err)
-		return err
-	}
-	proxyUrl, err := zedcloud.LookupProxy(
-		&ctx.deviceNetworkStatus, ifname, serverUrl)
-	if err == nil && proxyUrl != nil {
-		log.Infof("doHttp: Using proxy %s", proxyUrl.String())
-		dEndPoint.WithSrcIpAndProxySelection(ipSrc, proxyUrl)
-	} else {
-		dEndPoint.WithSrcIpSelection(ipSrc)
-	}
-	var respChan = make(chan *zedUpload.DronaRequest)
-
-	log.Infof("doHttp syncOp for <%s>, <%s>, <%s>\n", serverUrl, dpath,
-		filename)
-	// create Request
-	// Round up from bytes to Mbytes
-	maxMB := (maxsize + 1024*1024 - 1) / (1024 * 1024)
-	req := dEndPoint.NewRequest(syncOp, filename, locFilename, int64(maxMB), true, respChan)
-	if req == nil {
-		return errors.New("NewRequest failed")
-	}
-
-	req.Post()
-	for {
-		select {
-		case resp, ok := <-respChan:
-			if resp.IsDnUpdate() {
-				asize := resp.GetAsize()
-				osize := resp.GetOsize()
-				log.Infof("Update progress for %v: %v/%v",
-					resp.GetLocalName(), asize, osize)
-				if osize == 0 {
-					status.Progress = 0
-				} else {
-					percent := 100 * asize / osize
-					status.Progress = uint(percent)
-				}
-				publishDownloaderStatus(ctx, status)
-				continue
-			}
-			if !ok {
-				errStr := fmt.Sprintf("respChan EOF for <%s>, <%s>",
-					serverUrl, filename)
-				log.Errorln(errStr)
-				return errors.New(errStr)
-			}
-			if syncOp == zedUpload.SyncOpDownload {
-				err = resp.GetDnStatus()
-			} else {
-				_, err = resp.GetUpStatus()
-			}
-			if resp.IsError() {
-				return err
-			} else {
-				log.Infof("Done for %v: size %v/%v",
-					resp.GetLocalName(),
-					resp.GetAsize(), resp.GetOsize())
-				status.Progress = 100
-				publishDownloaderStatus(ctx, status)
-				return nil
-			}
-		}
-	}
-}
-
-func doS3(ctx *downloaderContext, status *types.DownloaderStatus,
-	syncOp zedUpload.SyncOpType, dnldUrl string, apiKey string, password string,
-	dpath string, region string, maxsize uint64, ifname string,
-	ipSrc net.IP, filename string, locFilename string) error {
-
-	auth := &zedUpload.AuthInput{
-		AuthType: "s3",
-		Uname:    apiKey,
-		Password: password,
-	}
-
-	trType := zedUpload.SyncAwsTr
-
-	// create Endpoint
-	dEndPoint, err := ctx.dCtx.NewSyncerDest(trType, region, dpath, auth)
-	if err != nil {
-		log.Errorf("NewSyncerDest failed: %s\n", err)
-		return err
-	}
-	// check for proxies on the selected management port interface
-	proxyUrl, err := zedcloud.LookupProxy(
-		&ctx.deviceNetworkStatus, ifname, dnldUrl)
-	if err == nil && proxyUrl != nil {
-		log.Infof("doS3: Using proxy %s", proxyUrl.String())
-		dEndPoint.WithSrcIpAndProxySelection(ipSrc, proxyUrl)
-	} else {
-		dEndPoint.WithSrcIpSelection(ipSrc)
-	}
-
-	var respChan = make(chan *zedUpload.DronaRequest)
-
-	log.Infof("doS3 syncOp for <%s>, <%s>, <%s>\n", dpath, region, filename)
-	// create Request
-	// Round up from bytes to Mbytes
-	maxMB := (maxsize + 1024*1024 - 1) / (1024 * 1024)
-	req := dEndPoint.NewRequest(syncOp, filename, locFilename,
-		int64(maxMB), true, respChan)
-	if req == nil {
-		return errors.New("NewRequest failed")
-	}
-
-	req.Post()
-	for {
-		select {
-		case resp, ok := <-respChan:
-			if resp.IsDnUpdate() {
-				asize := resp.GetAsize()
-				osize := resp.GetOsize()
-				log.Infof("Update progress for %v: %v/%v",
-					resp.GetLocalName(), asize, osize)
-				if osize == 0 {
-					status.Progress = 0
-				} else {
-					percent := 100 * asize / osize
-					status.Progress = uint(percent)
-				}
-				publishDownloaderStatus(ctx, status)
-				continue
-			}
-			if !ok {
-				errStr := fmt.Sprintf("respChan EOF for <%s>, <%s>, <%s>",
-					dpath, region, filename)
-				log.Errorln(errStr)
-				return errors.New(errStr)
-			}
-			if syncOp == zedUpload.SyncOpDownload {
-				err = resp.GetDnStatus()
-			} else {
-				_, err = resp.GetUpStatus()
-			}
-			if resp.IsError() {
-				return err
-			} else {
-				log.Infof("Done for %v: size %v/%v",
-					resp.GetLocalName(),
-					resp.GetAsize(), resp.GetOsize())
-				status.Progress = 100
-				publishDownloaderStatus(ctx, status)
-				return nil
-			}
-		}
-	}
-}
-
-func doAzureBlob(ctx *downloaderContext, status *types.DownloaderStatus,
-	syncOp zedUpload.SyncOpType, dnldURL string, apiKey string, password string,
-	dpath string, maxsize uint64, ifname string,
-	ipSrc net.IP, filename string, locFilename string) error {
-
-	auth := &zedUpload.AuthInput{
-		AuthType: "password",
-		Uname:    apiKey,
-		Password: password,
-	}
-
-	trType := zedUpload.SyncAzureTr
-
-	// create Endpoint
-	dEndPoint, err := ctx.dCtx.NewSyncerDest(trType, "", dpath, auth)
-	if err != nil {
-		log.Errorf("NewSyncerDest failed: %s\n", err)
-		return err
-	}
-	// check for proxies on the selected management port interface
-	proxyURL, err := zedcloud.LookupProxy(
-		&ctx.deviceNetworkStatus, ifname, dnldURL)
-	if err == nil && proxyURL != nil {
-		log.Infof("doAzure: Using proxy %s", proxyURL.String())
-		dEndPoint.WithSrcIpAndProxySelection(ipSrc, proxyURL)
-	} else {
-		dEndPoint.WithSrcIpSelection(ipSrc)
-	}
-
-	var respChan = make(chan *zedUpload.DronaRequest)
-
-	log.Infof("doAzure syncOp for <%s>, <%s>\n", dpath, filename)
-	// create Request
-	// Round up from bytes to Mbytes
-	maxMB := (maxsize + 1024*1024 - 1) / (1024 * 1024)
-	req := dEndPoint.NewRequest(syncOp, filename, locFilename,
-		int64(maxMB), true, respChan)
-	if req == nil {
-		return errors.New("NewRequest failed")
-	}
-
-	req.Post()
-	for {
-		select {
-		case resp, ok := <-respChan:
-			if resp.IsDnUpdate() {
-				asize := resp.GetAsize()
-				osize := resp.GetOsize()
-				log.Infof("Update progress for %v: %v/%v",
-					resp.GetLocalName(), asize, osize)
-				if osize == 0 {
-					status.Progress = 0
-				} else {
-					percent := 100 * asize / osize
-					status.Progress = uint(percent)
-				}
-				publishDownloaderStatus(ctx, status)
-				continue
-			}
-			if !ok {
-				errStr := fmt.Sprintf("respChan EOF for <%s>, <%s>",
-					dpath, filename)
-				log.Errorln(errStr)
-				return errors.New(errStr)
-			}
-			if syncOp == zedUpload.SyncOpDownload {
-				err = resp.GetDnStatus()
-			} else {
-				_, err = resp.GetUpStatus()
-			}
-			if resp.IsError() {
-				return err
-			} else {
-				log.Infof("Done for %v: size %v/%v",
-					resp.GetLocalName(),
-					resp.GetAsize(), resp.GetOsize())
-				status.Progress = 100
-				publishDownloaderStatus(ctx, status)
-				return nil
-			}
-		}
-	}
-}
-
-func doSftp(ctx *downloaderContext, status *types.DownloaderStatus,
-	syncOp zedUpload.SyncOpType, apiKey string, password string,
-	serverUrl string, dpath string, maxsize uint64,
-	ipSrc net.IP, filename string, locFilename string) error {
-
-	auth := &zedUpload.AuthInput{
-		AuthType: "sftp",
-		Uname:    apiKey,
-		Password: password,
-	}
-
-	trType := zedUpload.SyncSftpTr
-
-	// create Endpoint
-	dEndPoint, err := ctx.dCtx.NewSyncerDest(trType, serverUrl, dpath, auth)
-	if err != nil {
-		log.Errorf("NewSyncerDest failed: %s\n", err)
-		return err
-	}
-	dEndPoint.WithSrcIpSelection(ipSrc)
-	var respChan = make(chan *zedUpload.DronaRequest)
-
-	log.Infof("doSftp syncOp for <%s>, <%s>, <%s>\n", serverUrl, dpath,
-		filename)
-	// create Request
-	// Round up from bytes to Mbytes
-	maxMB := (maxsize + 1024*1024 - 1) / (1024 * 1024)
-	req := dEndPoint.NewRequest(syncOp, filename, locFilename,
-		int64(maxMB), true, respChan)
-	if req == nil {
-		return errors.New("NewRequest failed")
-	}
-
-	req.Post()
-	for {
-		select {
-		case resp, ok := <-respChan:
-			if resp.IsDnUpdate() {
-				asize := resp.GetAsize()
-				osize := resp.GetOsize()
-				log.Infof("Update progress for %v: %v/%v",
-					resp.GetLocalName(), asize, osize)
-				if osize == 0 {
-					status.Progress = 0
-				} else {
-					percent := 100 * asize / osize
-					status.Progress = uint(percent)
-				}
-				publishDownloaderStatus(ctx, status)
-				continue
-			}
-			if !ok {
-				errStr := fmt.Sprintf("respChan EOF for <%s>, <%s>",
-					dpath, filename)
-				log.Errorln(errStr)
-				return errors.New(errStr)
-			}
-			_, err = resp.GetUpStatus()
-			if resp.IsError() {
-				return err
-			} else {
-				log.Infof("Done for %v: size %v/%v",
-					resp.GetLocalName(),
-					resp.GetAsize(), resp.GetOsize())
-				status.Progress = 100
-				publishDownloaderStatus(ctx, status)
-				return nil
-			}
-		}
-	}
-}
-
 func constructDatastoreContext(config types.DownloaderConfig, status *types.DownloaderStatus, dst *types.DatastoreConfig) *types.DatastoreContext {
 	dpath := dst.Dpath
 	if status.ObjType == types.CertObj {
@@ -1434,519 +1102,6 @@ func constructDatastoreContext(config types.DownloaderConfig, status *types.Down
 		Region:          dst.Region,
 	}
 	return &dsCtx
-}
-
-// getContainerRegistry will extract container registry and form download url
-// for rktFetch
-func getContainerRegistry(url string) (string, string, error) {
-	var urlReg, registryReg, downReg *regexp.Regexp
-	var registry, downURL string
-	var err error
-	urlReg, err = regexp.Compile(`^docker:\/\/(?:.*).io\/(?:.*)`)
-	if err != nil {
-		return "", "", err
-	}
-	registryReg, err = regexp.Compile(`(?:.*)(?:\.io)`)
-	if err != nil {
-		return "", "", err
-	}
-	downReg, err = regexp.Compile(`(?:.*)(?:\.io)\/`)
-	if err != nil {
-		return "", "", err
-	}
-	if urlReg.MatchString(url) {
-		registry = strings.TrimPrefix(registryReg.FindString(url), "docker://")
-		downURL = downReg.ReplaceAllString(url, "docker://")
-		return registry, downURL, nil
-	}
-	return "", "", fmt.Errorf("Download URL is not formed properly")
-}
-
-func rktFetch(url string, localConfigDir string, pullPolicy string) (string, error) {
-
-	// rkt --insecure-options=image fetch <url> --dir=/persist/rkt --full=true
-	log.Debugf("rktFetch - url: %s ,  localConfigDir:%s\n",
-		url, localConfigDir)
-	cmd := "rkt"
-	args := []string{
-		"--dir=" + types.PersistRktDataDir,
-		"--insecure-options=image",
-		"fetch",
-	}
-	if len(localConfigDir) > 0 {
-		args = append(args, "--local-config="+localConfigDir)
-	}
-	if len(pullPolicy) > 0 {
-		args = append(args, "--pull-policy="+pullPolicy)
-	}
-	args = append(args, url)
-	args = append(args, "--full=true")
-
-	log.Infof("rktFetch - url: %s ,  localConfigDir:%s, args: %+v\n",
-		url, localConfigDir, args)
-
-	stdoutStderr, err := wrap.Command(cmd, args...).CombinedOutput()
-	if err != nil {
-		log.Errorln("rkt fetch failed ", err)
-		log.Errorln("rkt fetch output ", string(stdoutStderr))
-		return "", fmt.Errorf("rkt fetch failed: %s\nImage URL: %s\n",
-			string(stdoutStderr), url)
-	}
-	log.Infof("rktFetch - image fetch successful. stdoutStderr: %s\n",
-		stdoutStderr)
-	outputStr := string(stdoutStderr)
-	log.Debugf("rktFetch - outputStr: %s\n", outputStr)
-	outputStrArray := strings.Split(outputStr, "\n")
-
-	log.Debugf("rktFetch - outputStrArray:\n")
-	for i, op := range outputStrArray {
-		log.Debugf("index:%d, op:%s", i, op)
-	}
-	log.Debugf("rktFetch - outputStrArray DONE\n")
-
-	// Get ImageID from the oputput. The last line in rkt fetch output
-	// with sha12- is the imageID
-	imageID := ""
-	for i := len(outputStrArray) - 1; i >= 0; i-- {
-		imageID = outputStrArray[i]
-		if strings.HasPrefix(imageID, "sha512-") {
-			break
-		}
-	}
-	log.Infof("rktFetch - imageID: %s\n", imageID)
-	if imageID == "" {
-		errMsg := "rkt fetch: Can't find imageID.\n Fetch Output: " +
-			outputStr
-		return "", errors.New(errMsg)
-	}
-
-	// XXX:FIXME - we should run "rkt image ls" and verify image fetch
-	// went thru without errors.
-	return imageID, nil
-}
-
-// createRktLocalDirAndAuthFile
-//  Return Values: localConfigDir, AuthFileName, err
-func createRktLocalDirAndAuthFile(imageID uuid.UUID,
-	dsCtx types.DatastoreContext, registry string) (string, string, error) {
-
-	if len(strings.TrimSpace(dsCtx.APIKey)) == 0 {
-		log.Debugf("createRktLocalDirAndAuthFile: empty APIKey. " +
-			"Skipping AuthFile")
-		return "", "", nil
-	}
-
-	// Create Local Directory with name imageSafeName
-	// The directory structure should be:
-	// <persistRktLocalConfigBase>/<uuid>/auth.d/rktAuth<appName>.json
-	localConfigDir := persistRktLocalConfigBase + "/" + imageID.String()
-	authDir := localConfigDir + "/auth.d"
-	authFileName := authDir + "/docker.json"
-	err := os.MkdirAll(authDir, 0755)
-	if err != nil {
-		log.Errorf("createRktLocalDirAndAuthFile: empty username." +
-			" Skipping AuthFile")
-		return "", "", fmt.Errorf("Failed create dir %s, err: %+v\n",
-			authDir, err)
-	}
-
-	rktAuth := types.RktAuthInfo{
-		RktKind:    "dockerAuth",
-		RktVersion: "v1",
-		Registries: []string{registry},
-		Credentials: &types.RktCredentials{
-			User:     dsCtx.APIKey,
-			Password: dsCtx.Password,
-		},
-	}
-	log.Infof("createRktLocalDirAndAuthFile: created Auth file %s\n"+
-		"RktKind: %s, RktVersion: %s, Registries: %+v, \n",
-		authFileName, rktAuth.RktKind, rktAuth.RktVersion, rktAuth.Registries)
-
-	rktAuthJSON, err := json.MarshalIndent(rktAuth, "", " ")
-	if err != nil {
-		return "", "", fmt.Errorf("Failed convert rktAuth to json"+
-			"err: %+v\n", err)
-	}
-	err = ioutil.WriteFile(authFileName, rktAuthJSON, 0644)
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to create Auth file for"+
-			"rkt fetch: %+v\n", err)
-	}
-	return localConfigDir, authFileName, nil
-}
-
-func rktFetchContainerImage(ctx *downloaderContext, key string,
-	config types.DownloaderConfig, status *types.DownloaderStatus,
-	dsCtx types.DatastoreContext, pullPolicy string) error {
-	// update status to DOWNLOAD STARTED
-	status.State = types.DOWNLOAD_STARTED
-	publishDownloaderStatus(ctx, status)
-
-	imageID := ""
-	var authFile, localConfigDir string
-	registry, downloadURL, err := getContainerRegistry(dsCtx.DownloadURL)
-	if err == nil {
-		// Save credentials to Auth file
-		log.Infof("rktFetchContainerImage: fetch  <%s>\n", dsCtx.DownloadURL)
-		localConfigDir, authFile, err = createRktLocalDirAndAuthFile(
-			config.ImageID, dsCtx, registry)
-		if err == nil {
-			log.Debugf("rktFetchContainerImage: localConfigDir: %s, authFile: %s\n",
-				localConfigDir, authFile)
-			if len(authFile) == 0 {
-				log.Infof("rktFetchContainerImage: no Auth File")
-			}
-			imageID, err = rktFetch(downloadURL, localConfigDir, pullPolicy)
-		} else {
-			log.Errorf("rktCreateAuthFile Failed. err: %+v", err)
-		}
-	} else {
-		log.Errorf("getContainerRegistry : registry and download url parsing failed. err: %+v", err)
-	}
-
-	if err != nil {
-		log.Errorf("rktFetchContainerImage: fetch  Failed. url:%s, "+
-			"authFile: %s, Err: %+v\n", dsCtx.DownloadURL, authFile, err)
-		status.PendingAdd = false
-		status.Size = 0
-		status.LastErr = fmt.Sprintf("%v", err)
-		status.LastErrTime = time.Now()
-		status.RetryCount++
-		publishDownloaderStatus(ctx, status)
-		return err
-	}
-	log.Infof("rktFetchContainerImage successful. imageID: <%s>\n",
-		imageID)
-
-	// Update globalStatus and status
-	unreserveSpace(ctx, status)
-
-	// We do not clear any status.RetryCount, LastErr, etc. The caller
-	// should look at State == DOWNLOADED to determine it is done.
-	status.ContainerImageID = imageID
-	status.ContainerRktLocalConfigDir = localConfigDir
-	status.ContainerRktAuthFileName = authFile
-	status.ModTime = time.Now()
-	status.PendingAdd = false
-	status.State = types.DOWNLOADED
-	status.Progress = 100 // Just in case
-	publishDownloaderStatus(ctx, status)
-
-	return nil
-}
-
-// Drona APIs for object Download
-func handleSyncOp(ctx *downloaderContext, key string,
-	config types.DownloaderConfig, status *types.DownloaderStatus,
-	dst *types.DatastoreConfig) {
-	var err error
-	var errStr string
-	var locFilename string
-	var syncOp zedUpload.SyncOpType = zedUpload.SyncOpDownload
-	var serverURL string
-
-	if status.ObjType == "" {
-		log.Fatalf("handleSyncOp: No ObjType for %s\n",
-			status.Safename)
-	}
-
-	log.Debugf("handleSyncOp: config: %+v", config)
-	log.Debugf("handleSyncOp: IsContainer: %v", config.IsContainer)
-
-	// get the datastore context
-	dsCtx := constructDatastoreContext(config, status, dst)
-
-	if config.IsContainer {
-		pullPolicy := "new"
-		rktFetchContainerImage(ctx, key, config, status, *dsCtx, pullPolicy)
-		return
-	}
-	locDirname := types.DownloadDirname + "/" + status.ObjType
-	locFilename = locDirname + "/pending"
-
-	// update status to DOWNLOAD STARTED
-	status.State = types.DOWNLOAD_STARTED
-	publishDownloaderStatus(ctx, status)
-
-	if config.ImageSha256 != "" {
-		locFilename = locFilename + "/" + config.ImageSha256
-	}
-
-	if _, err := os.Stat(locFilename); err != nil {
-		log.Debugf("Create %s\n", locFilename)
-		if err = os.MkdirAll(locFilename, 0755); err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	filename := types.SafenameToFilename(config.Safename)
-
-	locFilename = locFilename + "/" + config.Safename
-
-	log.Infof("Downloading <%s> to <%s> using %v allow non-free port\n",
-		config.Name, locFilename, config.AllowNonFreePort)
-
-	var addrCount int
-	if !config.AllowNonFreePort {
-		addrCount = types.CountLocalAddrFreeNoLinkLocal(ctx.deviceNetworkStatus)
-		log.Infof("Have %d free management port addresses\n", addrCount)
-		err = errors.New("No free IP management port addresses for download")
-	} else {
-		addrCount = types.CountLocalAddrAnyNoLinkLocal(ctx.deviceNetworkStatus)
-		log.Infof("Have %d any management port addresses\n", addrCount)
-		err = errors.New("No IP management port addresses for download")
-	}
-	if addrCount == 0 {
-		errStr = err.Error()
-	}
-	metricsUrl := dsCtx.DownloadURL
-	if dsCtx.TransportMethod == zconfig.DsType_DsS3.String() {
-		// fake URL for metrics
-		metricsUrl = fmt.Sprintf("S3:%s/%s", dsCtx.Dpath, filename)
-	}
-
-	// Loop through all interfaces until a success
-	for addrIndex := 0; addrIndex < addrCount; addrIndex += 1 {
-		var ipSrc net.IP
-		if !config.AllowNonFreePort {
-			ipSrc, err = types.GetLocalAddrFreeNoLinkLocal(ctx.deviceNetworkStatus,
-				addrIndex, "")
-		} else {
-			// Note that GetLocalAddrAny has the free ones first
-			ipSrc, err = types.GetLocalAddrAnyNoLinkLocal(ctx.deviceNetworkStatus,
-				addrIndex, "")
-		}
-		if err != nil {
-			log.Errorf("GetLocalAddr failed: %s\n", err)
-			errStr = errStr + "\n" + err.Error()
-			continue
-		}
-		ifname := types.GetMgmtPortFromAddr(ctx.deviceNetworkStatus, ipSrc)
-		log.Infof("Using IP source %v if %s transport %v\n",
-			ipSrc, ifname, dsCtx.TransportMethod)
-		switch dsCtx.TransportMethod {
-		case zconfig.DsType_DsS3.String():
-			err = doS3(ctx, status, syncOp, dsCtx.DownloadURL, dsCtx.APIKey,
-				dsCtx.Password, dsCtx.Dpath, dsCtx.Region,
-				config.Size, ifname, ipSrc, filename, locFilename)
-			if err != nil {
-				log.Errorf("Source IP %s failed: %s\n",
-					ipSrc.String(), err)
-				errStr = errStr + "\n" + err.Error()
-				// XXX don't know how much we downloaded!
-				// Could have failed half-way. Using zero.
-				zedcloud.ZedCloudFailure(ifname,
-					metricsUrl, 1024, 0)
-			} else {
-				// Record how much we downloaded
-				size := int64(0)
-				info, err := os.Stat(locFilename)
-				if err != nil {
-					log.Error(err)
-				} else {
-					size = info.Size()
-				}
-				zedcloud.ZedCloudSuccess(ifname,
-					metricsUrl, 1024, size)
-				handleSyncOpResponse(ctx, config, status,
-					locFilename, key, "")
-				return
-			}
-		case zconfig.DsType_DsAzureBlob.String():
-			// pass in the config.Name instead of 'filename' which
-			// does not contain the prefix of the relative path with '/'s
-			err = doAzureBlob(ctx, status, syncOp, dsCtx.DownloadURL, dsCtx.APIKey,
-				dsCtx.Password, dsCtx.Dpath, config.Size, ifname, ipSrc, config.Name, locFilename)
-			if err != nil {
-				log.Errorf("Source IP %s failed: %s\n",
-					ipSrc.String(), err)
-				errStr = errStr + "\n" + err.Error()
-				// XXX don't know how much we downloaded!
-				// Could have failed half-way. Using zero.
-				zedcloud.ZedCloudFailure(ifname,
-					metricsUrl, 1024, 0)
-			} else {
-				// Record how much we downloaded
-				size := int64(0)
-				info, err := os.Stat(locFilename)
-				if err != nil {
-					log.Error(err)
-				} else {
-					size = info.Size()
-				}
-				zedcloud.ZedCloudSuccess(ifname,
-					metricsUrl, 1024, size)
-				handleSyncOpResponse(ctx, config, status,
-					locFilename, key, "")
-				return
-			}
-		case zconfig.DsType_DsSFTP.String():
-			serverURL, err = getServerUrl(dsCtx)
-			if err == nil {
-				// pass in the config.Name instead of 'filename' which
-				// does not contain the prefix of the relative path with '/'s
-				err = doSftp(ctx, status, syncOp, dsCtx.APIKey,
-					dsCtx.Password, serverURL, dsCtx.Dpath,
-					config.Size, ipSrc, config.Name, locFilename)
-			}
-			if err != nil {
-				log.Errorf("Source IP %s failed: %s\n",
-					ipSrc.String(), err)
-				errStr = errStr + "\n" + err.Error()
-				// XXX don't know how much we downloaded!
-				// Could have failed half-way. Using zero.
-				zedcloud.ZedCloudFailure(ifname,
-					metricsUrl, 1024, 0)
-			} else {
-				// Record how much we downloaded
-				size := int64(0)
-				info, err := os.Stat(locFilename)
-				if err != nil {
-					log.Error(err)
-				} else {
-					size = info.Size()
-				}
-				zedcloud.ZedCloudSuccess(ifname,
-					metricsUrl, 1024, size)
-				handleSyncOpResponse(ctx, config, status,
-					locFilename, key, "")
-				return
-			}
-		case zconfig.DsType_DsHttp.String(), zconfig.DsType_DsHttps.String(), "":
-			serverURL, err = getServerUrl(dsCtx)
-			if err == nil {
-				// pass in the config.Name instead of 'filename' which
-				// does not contain the prefix of the relative path with '/'s
-				err = doHttp(ctx, status, syncOp, serverURL, dsCtx.Dpath,
-					config.Size, ifname, ipSrc, config.Name, locFilename)
-			}
-			if err != nil {
-				log.Errorf("Source IP %s failed: %s\n",
-					ipSrc.String(), err)
-				errStr = errStr + "\n" + err.Error()
-				zedcloud.ZedCloudFailure(ifname,
-					metricsUrl, 1024, 0)
-			} else {
-				// Record how much we downloaded
-				size := int64(0)
-				info, err := os.Stat(locFilename)
-				if err != nil {
-					log.Error(err)
-				} else {
-					size = info.Size()
-				}
-				zedcloud.ZedCloudSuccess(ifname,
-					metricsUrl, 1024, size)
-				handleSyncOpResponse(ctx, config, status,
-					locFilename, key, "")
-				return
-			}
-		default:
-			errStr = "unsupported transport method " + dsCtx.TransportMethod
-		}
-	}
-	log.Errorf("All source IP addresses failed. All errors:%s\n", errStr)
-	handleSyncOpResponse(ctx, config, status, locFilename,
-		key, errStr)
-}
-
-// DownloadURL format : http://<serverURL>/dpath/filename
-func getServerUrl(dsCtx *types.DatastoreContext) (string, error) {
-	u, err := url.Parse(dsCtx.DownloadURL)
-	if err != nil {
-		log.Errorf("URL Parsing failed: %s\n", err)
-		return "", err
-	}
-	return u.Scheme + "://" + u.Host, nil
-}
-
-func handleSyncOpResponse(ctx *downloaderContext, config types.DownloaderConfig,
-	status *types.DownloaderStatus, locFilename string,
-	key string, errStr string) {
-
-	if status.ObjType == "" {
-		log.Fatalf("handleSyncOpResponse: No ObjType for %s\n",
-			status.Safename)
-	}
-	locDirname := types.DownloadDirname + "/" + status.ObjType
-	if errStr != "" {
-		// Delete file
-		doDelete(ctx, key, locDirname, status)
-		status.PendingAdd = false
-		status.Size = 0
-		status.LastErr = errStr
-		status.LastErrTime = time.Now()
-		status.RetryCount += 1
-		publishDownloaderStatus(ctx, status)
-		log.Errorf("handleSyncOpResponse failed for %s, <%s>\n",
-			status.Name, errStr)
-		return
-	}
-
-	info, err := os.Stat(locFilename)
-	if err != nil {
-		log.Errorf("handleSyncOpResponse Stat failed for %s <%s>\n",
-			status.Name, err)
-		// Delete file
-		doDelete(ctx, key, locDirname, status)
-		status.PendingAdd = false
-		status.Size = 0
-		status.LastErr = fmt.Sprintf("%v", err)
-		status.LastErrTime = time.Now()
-		status.RetryCount += 1
-		publishDownloaderStatus(ctx, status)
-		return
-	}
-	status.Size = uint64(info.Size())
-
-	// Update globalStatus and status
-	unreserveSpace(ctx, status)
-
-	log.Infof("handleSyncOpResponse successful <%s> <%s>\n",
-		config.Name, locFilename)
-	// We do not clear any status.RetryCount, LastErr, etc. The caller
-	// should look at State == DOWNLOADED to determine it is done.
-
-	status.ModTime = time.Now()
-	status.PendingAdd = false
-	status.State = types.DOWNLOADED
-	status.Progress = 100 // Just in case
-	publishDownloaderStatus(ctx, status)
-}
-
-// Handles both create and modify events
-func handleDNSModify(ctxArg interface{}, key string, statusArg interface{}) {
-
-	ctx := ctxArg.(*downloaderContext)
-	status := cast.CastDeviceNetworkStatus(statusArg)
-	if key != "global" {
-		log.Infof("handleDNSModify: ignoring %s\n", key)
-		return
-	}
-	log.Infof("handleDNSModify for %s\n", key)
-	if cmp.Equal(ctx.deviceNetworkStatus, status) {
-		log.Infof("handleDNSModify unchanged\n")
-		return
-	}
-	ctx.deviceNetworkStatus = status
-	log.Infof("handleDNSModify %d free management ports addresses; %d any\n",
-		types.CountLocalAddrFreeNoLinkLocal(ctx.deviceNetworkStatus),
-		types.CountLocalAddrAnyNoLinkLocal(ctx.deviceNetworkStatus))
-
-	log.Infof("handleDNSModify done for %s\n", key)
-}
-
-func handleDNSDelete(ctxArg interface{}, key string, statusArg interface{}) {
-
-	ctx := ctxArg.(*downloaderContext)
-	log.Infof("handleDNSDelete for %s\n", key)
-	if key != "global" {
-		log.Infof("handleDNSDelete: ignoring %s\n", key)
-		return
-	}
-	ctx.deviceNetworkStatus = types.DeviceNetworkStatus{}
-	log.Infof("handleDNSDelete done for %s\n", key)
 }
 
 // Handles both create and modify events
@@ -2008,4 +1163,9 @@ func lookupDatastoreConfig(ctx *downloaderContext, dsID uuid.UUID,
 	log.Debugf("Found datastore(%s) for %s\n", dsID, name)
 	dst := cast.CastDatastoreConfig(cfg)
 	return &dst, ""
+}
+
+func sourceFailureError(ip, ifname, url string, err error) {
+	log.Errorf("Source IP %s failed: %s\n", ip, err)
+	zedcloud.ZedCloudFailure(ifname, url, 1024, 0)
 }
