+++ pkg/pillar/zedUpload/datastore_req.go
@@ -130,6 +130,21 @@ func (req *DronaRequest) GetOsize() int64 {
 	return req.objectSize
 }
 
+// Progress return download progress
+// returns asize (bytes), osize (bytes), progress (percent out of 100)
+func (req *DronaRequest) Progress() (int64, int64, uint) {
+	req.Lock()
+	defer req.Unlock()
+	asize := req.asize
+	osize := req.objectSize
+	progress := uint(0)
+	if osize != 0 {
+		percent := 100 * asize / osize
+		progress = uint(percent)
+	}
+	return asize, osize, progress
+}
+
 func (req *DronaRequest) GetImageList() []string {
 	req.Lock()
 	defer req.Unlock()
